--- src/app/(normal)/profile/components/NotFoundFallback.tsx
@@ -1,5 +1,0 @@
-import { t } from '@lingui/macro';
-
-export default function NotFoundFallback() {
-    return <div className=" mt-40 px-5 text-center text-2xl font-bold">{t`This profile doesn't exist`}</div>;
-}
