+++ src/helpers/resolveSocialMediaProviderName.ts
@@ -1,10 +1,13 @@
-import { createLookupTableResolver } from "@masknet/shared-base";
+import { createLookupTableResolver } from '@masknet/shared-base';
 
-import { SocialPlatform } from "@/constants/enum.js";
+import { SocialPlatform } from '@/constants/enum.js';
 
-export const resolveSocialPlatformName = createLookupTableResolver<SocialPlatform, string>({
-            [SocialPlatform.Lens]: 'Lens Protocol',
-            [SocialPlatform.Farcaster]: 'Farcaster',
-        }, (platform) => {
-            throw new Error(`Unknown social platform: ${platform}`)
-        })
+export const resolveSocialPlatformName = createLookupTableResolver<SocialPlatform, string>(
+    {
+        [SocialPlatform.Lens]: 'Lens Protocol',
+        [SocialPlatform.Farcaster]: 'Farcaster',
+    },
+    (platform) => {
+        throw new Error(`Unknown social platform: ${platform}`);
+    },
+);
